+++ reckoner/config.py
@@ -17,6 +17,7 @@
 
 import os
 import logging
+from os.path import dirname
 
 from .command_line_caller import call
 
@@ -66,6 +67,10 @@ class Config(object):
             self._config['archive'] = archive
 
         return self._config['archive']
+
+    @property
+    def course_base_directory(self):
+        return dirname(self.course_path) or None
 
     def __setattr__(self, name, value):
         object.__setattr__(self, name, value)